// web/src/app/core/models/human-resources/employee.ts
export interface Employee {
  businessEntityId?: number;
  fullName?: string;
  fullAddress?: string;
  firstName?: string;
  middleName?: string;
  lastName?: string;
  jobTitle?: string;
  birthDate?: any;
  shiftId?: number;
  departmentId?: number;
  departmentName?: string;
  departmentStartDate?: any;
  phoneNumber?: string;
  phoneNumberTypeId?: number;
  emailAddressId?: number;
  emailAddress?: string;
  addressId?: number;
  addressLine1?: string;
  addressLine2?: string;
  city?: string;
  postalCode?: string;
  stateProvinceId?: number;
  stateProvinceName?: string;
  countryRegionCode?: string;
  countryRegionName?: string;
}